requirements copy.txt
argon2-cffi
certifi
chardet
click
Django
django-request
Flask
Flask-SQLAlchemy

idna
itsdangerous
Jinja2
lxml
MarkupSafe
multitasking

public
python-dateutil
pytz
query-string

requests
six
SQLAlchemy
urllib3
urlopen
Werkzeug
wheel
gunicorn

